# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Personalized Learning Multi-Agent System API
  version: 0.1.0
  description: OpenAPI document covering the main endpoints of all planned microservices.
servers:
  - url: http://localhost
...
paths:
  /api/users/users:
    post:
      tags: [user-service]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
  /api/users/users/{user_id}:
    get:
      tags: [user-service]
      summary: Get user detail
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
  /api/users/users/{user_id}/profile:
    get:
      tags: [user-service]
      summary: Get user profile
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile detail
...
components:
  schemas:
    UserCreate:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
        email:
          type: string
          nullable: true
    UserRead:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        email:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...